Expose a summary of socket activity for Intra
These metrics are chosen to match Intra's current
set of socket metrics (except for retry-related
metrics, since retry is not yet implemented in
this Handler).

Notably, this change renames one of the two tun2socks
packages (renamed to "tunnel"), and exposes bindings for
additional packages.  This appears to be necessary
because gobind can only expose types from packages that
are explicitly exported.