Improve protection of DNS queries
Go on Android cannot determine the system's preferred DNS servers.
This is normally fine, because Resolver.PreferGo defaults to false,
but we need to set PreferGo to true when protection is enabled,
in order to protect the DNS query's UDP socket.

This change requires a Protector to indicate the list of DNS
servers, so that the Resolver can create a protected socket to
one of those servers.